Eoc exercises rename (#195)
* make BakeFirstElements more general & refactor solutions baking into BakeNumberedExercise (#191)

* Rename directions (#193)

* Rename directions

* bump to major
