+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.2.0] - 2021-04-19
 
+* Rename several directions (major)
+  * `BakeChapterReviewExercises` -> `MoveExercisesToEOC`
+  * `BakeChapterReview` -> `ChapterReviewContainer`
+  * `BakeBookAnswerKey` -> `BookAnswerKeyContainer`
+  * `BakeChapterAnswerKey` -> `MoveSolutionsToAnswerKey`
+* Refactors `BakeFirstElements` and `BakeNumberedExercise` (minor)
 * Adds a decorating option of equation numbers on `BakeEquations` (minor)
 
 ## [3.1.0] - 2021-04-19
