--- lib/kitchen/directions/bake_book_answer_key/v1.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Kitchen::Directions::BakeBookAnswerKey
-  class V1
-    renderable
-
-    def bake(book:)
-      @metadata = book.metadata.children_to_keep.copy
-      book.body.append(child: render(file: 'eob_solutions_container.xhtml.erb'))
-      book.body.first('.os-eob.os-solutions-container')
-    end
-  end
-end
