+++ lib/kitchen/directions/bake_numbered_exercise/v1.rb
@@ -9,17 +9,8 @@ module Kitchen::Directions::BakeNumberedExercise
       problem_number = "<span class='os-number'>#{number}</span>"
 
       if solution.present?
-        solution.id = "#{exercise.id}-solution"
-        exercise.add_class('os-hasSolution')
         problem_number = "<a class='os-number' href='##{exercise.id}-solution'>#{number}</a>"
-
-        solution.replace_children(with:
-          <<~HTML
-            <a class='os-number' href='##{exercise.id}'>#{number}</a>
-            <span class='os-divider'>. </span>
-            <div class="os-solution-container">#{solution.children}</div>
-          HTML
-        )
+        bake_solution(exercise: exercise, number: number)
       end
 
       problem.replace_children(with:
@@ -27,6 +18,20 @@ module Kitchen::Directions::BakeNumberedExercise
           #{problem_number}
           <span class='os-divider'>. </span>
           <div class="os-problem-container">#{problem.children}</div>
+        HTML
+      )
+    end
+
+    def bake_solution(exercise:, number:, divider: '. ')
+      solution = exercise.solution
+      solution.id = "#{exercise.id}-solution"
+      exercise.add_class('os-hasSolution')
+
+      solution.replace_children(with:
+        <<~HTML
+          <a class='os-number' href='##{exercise.id}'>#{number}</a>
+          <span class='os-divider'>#{divider}</span>
+          <div class="os-solution-container">#{solution.children}</div>
         HTML
       )
     end
