+++ spec/directions/move_solutions_to_answer_key/main_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+RSpec.describe Kitchen::Directions::MoveSolutionsToAnswerKey do
+  it 'calls v1' do
+    expect_any_instance_of(Kitchen::Directions::MoveSolutionsToAnswerKey::V1).to receive(:bake)
+      .with(chapter: 'chapter1', metadata_source: 'metadata', strategy: 'strategy', append_to: 'element')
+    described_class.v1(chapter: 'chapter1', metadata_source: 'metadata', strategy: 'strategy', append_to: 'element')
+  end
+end
